feat: faq router setting - add
