+++ src/component/qna/QNAReadComponent.jsx
@@ -1,0 +1,10 @@
+
+function QnaReadComponent() {
+    return (
+        <div>
+            <div>QNA Read Component</div>
+        </div>
+    );
+}
+
+export default QnaReadComponent;
