--- src/page/AdminDetailPage.jsx
@@ -1,9 +1,0 @@
-import AdminDetailComponent from "../common/AdminDetailComponent.jsx";
-
-function AdminDetailPage() {
-    return (
-        <AdminDetailComponent/>
-    );
-}
-
-export default AdminDetailPage;
