+++ src/router/CareTakerRouter.jsx
@@ -1,8 +1,8 @@
 import {lazy} from "react";
 
 
-const CareTakerIndex = lazy(() => import("../page/CareTakerIndexPage.jsx"))
-const CareTakerList = lazy(() => import("../page/CareTakerPage.jsx"))
+const CareTakerIndex = lazy(() => import("../page/CareTaker/CareTakerIndexPage.jsx"))
+const CareTakerList = lazy(() => import("../page/CareTaker/CareTakerPage.jsx"))
 
 const CareTakerRouter = {
